\name{publicAC}
\Rdversion{1.1}
\alias{publicA}
\alias{publicC}

\title{publicA and publicC}

\description{
  Two great function that do very simple mathematical operation.
}

\usage{
  publicA(x)
  publicC(x)
}

\arguments{
  \item{x}{\code{x} is a numeric}
}

\details{
  So simple, no details are needed.
}

\value{A numeric}

\references{Book: "Petit Manuel de Programmation Orientee Objet sous R"}

\author{Christophe Genolini <genolini@u-paris10.fr>}

\examples{
publicA(3)
publicC(4)
}
\keyword{documentation}
