\name{publicAC}
\alias{publicA}
\alias{publicC}

\title{publicA and publicC} % Toy functions for packClassic
\description{
Two great function that do very simple mathematical operation.
}
\usage{
publicA(x)
publicC(x)
}
\arguments{
  \item{x}{ \code{x} is a numeric}
}
\details{
So simple, no details are needed.
}
\value{A numeric}

\references{S4: From An Idea To Its Package}

\author{Christophe Genolini <genolini@u-paris10.fr>}

\examples{
publicA(3)
publicC(4)
}
\keyword{documentation}
